fixed execution permission

+++ dsi-runtime/samples/kafka/sample_test.sh
@@ -2,14 +2,21 @@
 
 set -e
 
-TMP_DIR="/tmp/sample-kafka-$$"
+if [ ! -z "$1" ]; then
+        TMP_DIR="/tmp/sample-kafka-$$"
 
-mkdir -p "$TMP_DIR"
-cd "$TMP_DIR"
+        mkdir -p "$TMP_DIR"
+        cd "$TMP_DIR"
 
-git clone https://github.com/ODMDev/odm-dsi-docker
+        git clone https://github.com/ODMDev/odm-dsi-docker
 
-cd dsi-runtime/samples/kafka/
+        cd "$TMP_DIR/odm-dsi-docker/dsi-runtime/samples/kafka/"
+fi
+
 ./start.sh
 
-rm -r "$TMP_DIR"
+docker-compose down -v
+
+if [ ! -z "$1" ]; then
+        rm -r "$TMP_DIR"
+fi
